+++ src/System/LoggerType.ts
@@ -1,0 +1,11 @@
+/**
+ * Logger type enumeration
+ *
+ * @enum
+ */
+enum LoggerType {
+  CONSOLE = "console",
+  FILE = "file",
+}
+
+export default LoggerType;